requirements.txt
arrow==0.17.0
beautifulsoup4==4.9.3
Django==3.0.2
#wasyprint
weasyprint
pydyf==0.2.0
html5lib==1.1
tinycss2==1.0.0
cssselect2==0.1
Pyphen==0.9.1
fontTools==4.0.0
#-----------
psycopg2-binary==2.8.6
django-ckeditor==5.9.0
django-import-export==2.5.0
django-js-asset==1.2.2
django-multiselectfield==0.1.12
importlib-metadata==1.6.1
importlib-resources==1.5.0
itsdangerous==1.1.0
Pillow==8.0.0
pytz==2021.1
reportlab==3.5.53
requests==2.25.1